Moved offline screen to design module
