+++ app/src/main/res/values/strings.xml
@@ -1,11 +1,6 @@
 <resources>
     <string name="app_name">GOV.UK</string>
 
-    <string name="device_offline_title">You are not connected to the internet</string>
-    <string name="device_offline_description_1">You need to have an internet connection to use the GOV.UK app.</string>
-    <string name="device_offline_description_2">Reconnect to the internet and try again.</string>
-    <string name="device_offline_button_title">Try again</string>
-
     <string name="forced_update_button_title_update">Update</string>
     <string name="forced_update_description">You’re using an old version of the app. Update your app to continue.</string>
     <string name="forced_update_title">You need to update your app</string>
